+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 20220623 [v2.1.1]
+
+- [FIXED] primevue adapter position bottom should convert to bottom-center
+- [IMPROVEMENT] refine types in primevue adapter
+
 ## 20220605 [v2.1.0]
 
 - [BREAKING] add adapter for convert default flashMessage object to primeVue format.